+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Change log
 - Added configuration to make running prettier and ESLint easy.
 - Added a different message when a region is empty and its ``inherited``
   flag is set.
+- Make the ``regions`` attribute on ``ContentEditorInline`` objects a
+  callable.
 
 
 `1.3`_ (2018-12-10)